[chore] #1 패키징 변경
