+++ app/src/main/java/com/sopt/now/advanced/team2android/SoptApp.kt
@@ -1,0 +1,7 @@
+package com.sopt.now.advanced.team2android
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class SoptApp : Application()
