--- app/src/main/java/com/sopt/now/advenced/team2android/SoptApp.kt
@@ -1,7 +1,0 @@
-package com.sopt.now.advenced.team2android
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class SoptApp : Application()
