+++ core/src/main/java/org/sopt/now/advanced/team2android/core/ui/base/BindingActivity.kt
@@ -1,0 +1,17 @@
+package org.sopt.now.advanced.team2android.core.ui.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import androidx.appcompat.app.AppCompatActivity
+import androidx.viewbinding.ViewBinding
+
+abstract class BindingActivity<T : ViewBinding>(private val inflater: (LayoutInflater) -> T) :
+    AppCompatActivity() {
+    lateinit var binding: T
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = inflater(layoutInflater)
+        setContentView(binding.root)
+    }
+}
