+++ core/src/main/java/org/sopt/now/advanced/team2android/core/ui/util/context/ContextExt.kt
@@ -1,0 +1,14 @@
+package org.sopt.now.advanced.team2android.core.ui.util.context
+
+import android.content.Context
+import androidx.annotation.ColorRes
+import androidx.annotation.StringRes
+import androidx.core.content.ContextCompat
+
+fun Context.stringOf(
+    @StringRes resId: Int
+) = getString(resId)
+
+fun Context.colorOf(
+    @ColorRes resId: Int
+) = ContextCompat.getColor(this, resId)
