+++ core/src/main/java/org/sopt/now/advanced/team2android/core/ui/util/fragment/FragmentExt.kt
@@ -1,0 +1,14 @@
+package org.sopt.now.advanced.team2android.core.ui.util.fragment
+
+import androidx.annotation.ColorRes
+import androidx.annotation.StringRes
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
+
+fun Fragment.stringOf(
+    @StringRes resId: Int
+) = getString(resId)
+
+fun Fragment.colorOf(
+    @ColorRes resId: Int
+) = ContextCompat.getColor(requireContext(), resId)
