+++ core/src/main/java/org/sopt/now/advanced/team2android/core/ui/util/intent/IntentExt.kt
@@ -1,0 +1,15 @@
+package org.sopt.now.advanced.team2android.core.ui.util.intent
+
+import android.content.Intent
+import android.os.Build
+import android.os.Parcelable
+
+inline fun <reified T : Parcelable> Intent.getCompatibleParcelableExtra(key: String): T? =
+    when {
+        Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU -> getParcelableExtra(
+            key,
+            T::class.java
+        )
+
+        else -> getParcelableExtra(key) as? T
+    }
