+++ core/src/main/java/org/sopt/now/advanced/team2android/core/ui/util/view/ItemDiffCallback.kt
@@ -1,0 +1,19 @@
+package org.sopt.now.advanced.team2android.core.ui.util.view
+
+import androidx.recyclerview.widget.DiffUtil
+
+class ItemDiffCallback<T : Any>(
+    val onItemsTheSame: (T, T) -> Boolean,
+    val onContentsTheSame: (T, T) -> Boolean
+) : DiffUtil.ItemCallback<T>() {
+    override fun areItemsTheSame(
+        oldItem: T,
+        newItem: T
+    ): Boolean = onItemsTheSame(oldItem, newItem)
+
+    override fun areContentsTheSame(
+        oldItem: T,
+        newItem: T
+    ): Boolean =
+        onContentsTheSame(oldItem, newItem)
+}
