--- core/src/main/java/org/sopt/now/advenced/team2android/core/ui/base/BindingBottomSheetDialogFragment.kt
@@ -1,28 +1,0 @@
-package org.sopt.now.advenced.team2android.core.ui.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.viewbinding.ViewBinding
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-
-abstract class BindingBottomSheetDialogFragment<T : ViewBinding>(private val inflater: (LayoutInflater) -> T) :
-    BottomSheetDialogFragment() {
-    private var _binding: T? = null
-    val binding get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다." }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = inflater(layoutInflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
-    }
-}
