--- core/src/main/java/org/sopt/now/advenced/team2android/core/ui/base/BindingFragment.kt
@@ -1,29 +1,0 @@
-package org.sopt.now.advenced.team2android.core.ui.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.viewbinding.ViewBinding
-
-abstract class BindingFragment<T : ViewBinding>(private val inflater: (LayoutInflater) -> T) :
-    Fragment() {
-    private var _binding: T? = null
-    private val binding
-        get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다." }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = inflater(layoutInflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
-    }
-}
