+++ data/src/main/java/com/sopt/now/advanced/team2android/data/MyClass.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.advanced.team2android.data
+
+class MyClass {
+}
