--- data/src/main/java/com/sopt/now/advenced/team2android/data/MyClass.kt
@@ -1,4 +1,0 @@
-package com.sopt.now.advenced.team2android.data
-
-class MyClass {
-}
