+++ domain/src/main/java/org/sopt/now/advanced/team2android/domain/entity/UserEntity.kt
@@ -1,0 +1,6 @@
+package org.sopt.now.advanced.team2android.domain.entity
+
+data class UserEntity(
+    val id: Int,
+    val firstName: String
+)
