--- domain/src/main/java/org/sopt/now/advenced/team2android/domain/entity/UserEntity.kt
@@ -1,6 +1,0 @@
-package org.sopt.now.advenced.team2android.domain.entity
-
-data class UserEntity(
-    val id: Int,
-    val firstName: String
-)
